fix: created different file to contain our different classes and exported it to our main.rb file
