+++ bin/main.rb
@@ -1,4 +1,6 @@
-require_relative '../lib/board.rb'
+require_relative '../lib/player.rb'
+require_relative '../lib/game.rb'
+require_relative '../lib/result.rb'
 
 module Umpire
     def self.display(arr)
